requirements.txt
# Generated on: Python 3.7.7
# With snowflake-connector-python version: 2.3.4
asn1crypto==1.4.0
auditwheel==3.1.1
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
boto3==1.15.18
botocore==1.18.18
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
cryptography==2.9.2
idna==2.10
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.19
oauthlib==3.1.0
oscrypto==1.2.1
pycparser==2.20
pycryptodomex==3.9.8
pyelftools==0.26
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.11
snowflake-connector-python==1.7.8
